ASE_project/Icebreaker/requirements1.txt
amqp==2.3.2
anyjson==0.3.3
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
bleach==2.1.3
braintree==3.49.0
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
cssutils==1.0.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-celery==3.2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-channels==0.7.0
django-ckeditor==5.6.1
django-common-helpers==0.9.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-cron==0.5.1
django-crontab==0.7.1
django-currentuser==0.2.3
django-filter==2.0.0
django-inlinecss==0.1.2
django-js-asset==1.1.0
django-liked==1.0.2
django-mathfilters==0.4.0
django-recaptcha==1.4.0
django-timezone-field==3.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
entrypoints==0.2.3
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.2.2
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.3
mock==2.0.0
mysql-connector-python==8.0.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.1
oauthlib==2.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.1
pickleshare==0.7.4
Pillow==5.3.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
Pygments==2.2.0
PyJWT==1.6.4
pynliner3==0.6
PySocks==1.6.8
python-crontab==2.3.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.3.1
requests==2.19.0
requests-oauthlib==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
stripe==1.82.1
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
twilio==6.19.2
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
